adding sweetalert confirm dialog before deleting product

+++ src/app/components/product-list/product-list.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Product } from 'src/app/interfaces/product';
 import { ProductService } from 'src/app/services/product.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-product-list',
@@ -34,11 +35,31 @@ export class ProductListComponent implements OnInit {
   }
 
   deleProductById(id: number) {
-    this.productService.deleteProductByID(id).subscribe({
-      next: () => {
-        this.listProduct();
+    Swal.fire({
+      title: "Are you sure that you want to delete this product?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!"
+    }).then((result) => {
+      if (result.isConfirmed) {
+
+        this.productService.deleteProductByID(id).subscribe({
+          next: () => {
+            this.listProduct();
+          }
+        });
+
+        Swal.fire({
+          title: "Deleted!",
+          text: "The product has been deleted.",
+          icon: "success"
+        });
       }
     });
+
   }
 
 }
